address pr feedback
